REF: libalgos tests
